renaming the input text view.
